+++ packages/image_picker/image_picker_platform_interface/lib/src/types/picked_file/base.dart
@@ -5,7 +5,7 @@
 import 'dart:convert';
 import 'dart:typed_data';
 
-import 'package:meta/meta.dart';
+import 'package:flutter/foundation.dart' show immutable;
 
 /// The interface for a PickedFile.
 ///