+++ packages/google_sign_in/google_sign_in_web/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.10.0+4
+
+* Removes dependency on `meta`.
+
 ## 0.10.0+3
 
 * Updated URL to the `google_sign_in` package in README.