+++ packages/path_provider/path_provider_macos/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.0.5
+
+* Removes dependency on `meta`.
+
+## 2.0.4
+
+* Switches to a package-internal implementation of the platform interface.
+
 ## 2.0.3
 
 * Fixes link in README.