+++ packages/webview_flutter/webview_flutter_platform_interface/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 1.8.1
+
+* Update to use the `verify` method introduced in platform_plugin_interface 2.1.0.
+
+## 1.8.0
+
+* Adds the `loadFlutterAsset` method to the platform interface.
+
+## 1.7.0
+
+* Add an option to set the background color of the webview.
+
 ## 1.6.1
 
 * Revert deprecation of `clearCookies` in WebViewPlatform for later deprecation.